Moved viewport-related structs and functionality to a new file.
