+++ crates/lgn-graphics-renderer/src/core/services/mod.rs
@@ -1,2 +1,5 @@
 mod cameras;
 pub(crate) use cameras::*;
+
+mod viewports;
+pub(crate) use viewports::*;
